Fix: Attempt to fix CI
